fix ffi-check on the free-threaded build
