+++ pyo3-ffi/src/cpython/object.rs
@@ -316,6 +316,8 @@ pub struct PyHeapTypeObject {
     ht_token: *mut c_void,
     #[cfg(all(Py_3_11, not(PyPy)))]
     _spec_cache: _specialization_cache,
+    #[cfg(Py_GIL_DISABLED)]
+    unique_id: Py_ssize_t,
 }
 
 impl Default for PyHeapTypeObject {
